First cut on vcb-guacamole-cluster script
Also a bunch of refactoring, and README updates.

Closes #11
