+++ README.md
@@ -1,6 +1,10 @@
 # variant-calling-benchmarks
 Automated and curated variant calling benchmarks for Guacamole
 
+This repository was created to aid development of the [Guacamole](https://github.com/hammerlab/guacamole) somatic variant caller. Eventually we may support benchmarking other variant callers.
+
+The sequencing data for most benchmarks comes from publically available datasets but cannot be directly released by us. To run these benchmarks you will have to acquire the BAMs yourself first.
+
 ## Installation
 
 From a checkout run:
@@ -12,13 +16,23 @@ Also checkout and build guacamole.
 
 ## Config files
 
-The benchmark defintions and guacamole configurations are given in JSON. They are split into multiple files for clarity but the 'configuration' is just the union of all the files passed, i.e. the code
-doesn't care what the files are or which file has a particular property.
+The benchmark defintions and guacamole configurations are given in JSON. They are split into multiple files for clarity but the 'configuration' is just the union of all the files passed. The code doesn't care what the files are or which file has a particular property.
 
 Paths in config files are relative to the directory of the config file. A simple substitution mechanism is supported where strings like "foo-${NAME}" will have $NAME expanded according to substitutions defined in the config files (under the 'substitutions' keys).
 
+## Benchmark definition
+
+The [local-wgs1 benchmark](benchmarks/local-wgs1/benchmark.json) is a small benchmark that can be run locally using BAMs that are checked into the repo.
+
+A benchmark consists of a set of *patients* and a set of *variants collections*. Each patient has at least two BAMs (tumor and normal) but can have more, including RNA. A variant collection is any set of variants we want to compare guacamole against, such as validated calls or calls generated by another caller. Variant collections are given as a single CSV file containing variants for all patients in the benchmark (see [here](benchmarks/local-wgs1/published_calls.csv) for an example).
+
+When a benchmark is run, sensitivity, specificity, and other metrics are computed across patients by comparing the guacamole calls to each of the variant collections. A CSV file of merged calls from guacamole and the variant collections is also written so further comparisons of these call sets may be performed. The guacamole invocation force calls the sites in the benchmark variant collections so guacamole likelihood and filter information is available for all sites in the benchmark whether or not they triggered a guacamole call.
+
+Note that variant calls are associated with patients, not samples. This is the "joint calling" approach: we do not call variants on individual BAMs but rather on the entire patient using all BAMs.
 
 ## Running Locally
+
+Change the `--guacamole-jar` path below for your environment.
 
 ```
 vcb-guacamole-local \
@@ -32,18 +46,21 @@ vcb-guacamole-local \
 
 ```
 
-## Running on demeter
+## Running on a cluster
 
-TODO, this does not work yet
+We override `TMPDIR` because the default (`/tmp`) is not mounted on our cluster nodes. The node you call the script from and the spark driver node (not the executors) must be able to share the temporary files. The driver node must also be able to write to the directory specified in `--out-dir`.
+
+Change the `--guacamole-jar` and `--guacamole-dependencies-jar` paths below for your environment.
 
 ```
-vcb-guacamole-demeter \
+mkdir tmp
+TMPDIR=$(pwd)/tmp vcb-guacamole-cluster \
     base_config.json \
-    infrastructure-configs/local.json \
+    infrastructure-configs/demeter.json \
     guacamole-configs/default.json \
     benchmarks/aocs/benchmark.json \
-    --guacamole-jar ~/sinai/git/guacamole/target/guacamole-with-dependencies-0.0.1-SNAPSHOT.jar \
+    --guacamole-jar ~/sinai/git/guacamole/target/guacamole-0.0.1-SNAPSHOT.jar \
+    --guacamole-dependencies-jar \
+        ~/sinai/git/guacamole/target/guacamole-deps-only-0.0.1-SNAPSHOT.jar \
     --out-dir results
-
-
 ```
