+++ benchmarks/aocs/benchmark.json
@@ -6,25 +6,26 @@
 
 "patients": {
     "AOCS-034": {
+        "loci": "$AUTOSOMAL_LOCI_WITH_CHR_PREFIXES",
         "reads": {
-            "normal blood": {
+            "normal_blood": {
                 "analyte": "dna",
-                "tissue": "normal",
+                "tissue_type": "normal",
                 "path": "$AOCS_HDFS_BAMS/EGAZ00001018680_2014_ICGC_IcgcOvarian_AOCS034_1DNA_1NormalBlood_ICGCDBPC20130205008_IlluminaIGNOutsourcing_NoCapture_Bwa_HiSeq.jpearson.header_fixed.bam"
             },
             "primary": {
                 "analyte": "dna",
-                "tissue": "tumor",
+                "tissue_type": "tumor",
                 "path": "$AOCS_HDFS_BAMS/EGAZ00001018597_2014_ICGC_IcgcOvarian_AOCS034_1DNA_7PrimaryTumour_ICGCDBPC20130205009_IlluminaIGNOutsourcing_NoCapture_Bwa_HiSeq.jpearson.header_fixed.bam"
             },
             "relapse": {
                 "analyte": "dna",
-                "tissue": "tumor",
+                "tissue_type": "tumor",
                 "path": "$AOCS_HDFS_BAMS/EGAZ00001018766_2014_ICGC_IcgcOvarian_AOCS034_1DNA_12Ascites_ICGCDBPC20130205007_IlluminaIGNOutsourcing_NoCapture_Bwa_HiSeq.jpearson.bam"
             }
-        },
-    },
-}
+        }
+    }
+},
 
 "variants": {
     "published": {
