+++ infrastructure-configs/local.json
@@ -4,10 +4,10 @@
     "AOCS_HDFS_BAMS": "hdfs:///datasets/ovarian/ega-tim"
 },
 
-"java_args": ["-Xmx4g", "-XX:MaxPermSize=512m"],
-
-"spark_configuration": {
-    "spark.master": "local[1]"
-}
+"java_arguments": [
+    "-Xmx4g",
+    "-XX:MaxPermSize=512m",
+    "-Dspark.master=local[1]"
+]
 
 }
