+++ variant_calling_benchmarks/guacamole_cluster.py
@@ -1,0 +1,61 @@
+'''
+Run guacamole on a benchmark using spark-submit.
+'''
+
+import sys
+import os
+import argparse
+import subprocess
+import logging
+
+from .config import load_config
+from . import temp_files
+from . import common
+from .joint_caller import invoke
+from .joint_caller import process_results
+
+parser = argparse.ArgumentParser(description=__doc__)
+common.add_common_run_args(parser)
+parser.add_argument("--guacamole-dependencies-jar",
+    required=True,
+    help="Path to guacamole-deps-only-VERSION.jar")
+parser.add_argument("--spark-submit", default="spark-submit",
+    help="Path to spark-submit script")
+
+def run(argv=sys.argv[1:]):
+    args = parser.parse_args(argv)
+    config = load_config(*args.configs)
+    try:
+        main(args, config)
+    finally:
+        temp_files.finished(not args.keep_temp_files)
+
+def main(args, config):
+    patients = args.patient if args.patient else sorted(config['patients'])
+
+    patient_to_vcf = {}
+
+    for patient in patients:
+        out_vcf = os.path.join(
+            os.path.abspath(args.out_dir),
+            "out.%s.%s.vcf" % (config['benchmark'], patient))
+        patient_to_vcf[patient] = out_vcf
+        logging.info("Running on patient %s outputting to %s" % (
+            patient, out_vcf))
+
+        invocation = (
+            [args.spark_submit] +
+            config["spark_submit_arguments"] + 
+            ["--jars", args.guacamole_dependencies_jar] +
+            ["--class", "org.hammerlab.guacamole.Main", args.guacamole_jar] +
+            invoke.make_arguments(config, patient, out_vcf))
+
+        if args.skip_guacamole:
+            logging.info("Skipping guacamole run with arguments %s" % str(
+                invocation))
+        else:
+            logging.info("Running guacamole with arguments %s" % str(
+                invocation))
+            subprocess.check_call(invocation)
+
+    process_results.write_merged_calls(args, config, patient_to_vcf)
